# [NDTensors] Reducing compilation time of dense tensor contraction