This PR is based on the discussion [here](https://itensor.discourse.group/t/strange-compilation-behavior-for-the-function-mps/391/15).

# Description

As summarized in the above page, ITensors.jl suffers from a long compilation time when decomposing a tensor into MPS. The compilation time increases as the number of indices is increased. Furthermore, even after naive code for a small number of indices is generated, the compilation time for a larger number of indices is not reduced. The same applies to reconstructing a tensor from an MPS.

I found that a substantial portion of the compilation time is spent on *compiling the code for analyzing contraction labels* in `contraction_logic.jl`. To address this issue with a minimal change, I fuse contraction labels before the actual contraction code is called.

Original labels:

```
(1, 2, -1, -2) * (-1, -2, 3, 4) =>  (1, 2, 3, 4)
```

After fusion:
```
(1, 2) * (2, 3) =>  (1, 3)
```

This fusing trick substantially reduces the number of concrete types for `ContractionProperties{NA,NB,NC}`.
The following benchmarks were performed on Macbook Pro 14 (M1) with a single thread.

## Benchmark 1

```Julia
using ITensors
ITensors.disable_warn_order()

function run1(ElemType::Type{T}, nbit) where {T}
    sites = siteinds("Qubit", nbit)
    M = randomMPS(sites)
    fulltensor = reduce(*, M)
    return nothing
end

for nbit in [4, 8, 16, 24]
    @time run1(Float64, nbit)
end
```

v0.3.24:
```
7.673702 seconds (48.42 M allocations: 2.709 GiB, 8.71% gc time, 99.98% compilation time)
10.310319 seconds (51.52 M allocations: 3.140 GiB, 6.30% gc time, 99.99% compilation time)
21.941443 seconds (111.89 M allocations: 6.543 GiB, 6.95% gc time, 99.99% compilation time)
26.910477 seconds (134.17 M allocations: 7.751 GiB, 6.21% gc time, 99.70% compilation time)
```

New code:
```
 4.346442 seconds (28.34 M allocations: 1.551 GiB, 9.49% gc time, 99.96% compilation time)
 0.753839 seconds (4.43 M allocations: 229.441 MiB, 4.92% gc time, 99.85% compilation time)
 2.053619 seconds (11.85 M allocations: 595.432 MiB, 4.50% gc time, 99.90% compilation time)
 3.446269 seconds (17.43 M allocations: 1.075 GiB, 5.84% gc time, 96.90% compilation time)
```

## Benchmark 2

```Julia
using Pkg
@time Pkg.test("ITensors")
```

v0.3.24:

* 1167.520032 seconds
* 1178.514122 seconds

New code:

* 973.487310 seconds
* 977.637775 seconds


# How Has This Been Tested?

- [x] [Fuse label](https://github.com/shinaoka/ITensors.jl/blob/contract_opt2/NDTensors/test/dense.jl#L260) Check if labels are fused correctly
- [x] [contract with fusion](https://github.com/shinaoka/ITensors.jl/blob/contract_opt2/NDTensors/test/dense.jl#L271) Check if `contract!` returns the correct value with fusion.
- [x] All other tests in `NDTensors.jl` and `ITensors.jl` pass.

# Checklist:

- [x] My code follows the style guidelines of this project. Please run `using JuliaFormatter; format(".")` in the base directory of the repository (`~/.julia/dev/ITensors`) to format your code according to our style guidelines.
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [x] I have added tests that verify the behavior of the changes I made.
- [ ] I have made corresponding changes to the documentation. (No change in behavior!)
- [x] My changes generate no new warnings.
- [x] Any dependent changes have been merged and published in downstream modules.