@shinaoka 2023-02-17
There is one error in `ITensorGPU.jl`:

https://jenkins.flatironinstitute.org/blue/organizations/jenkins/ITensors.jl/detail/PR-1060/2/pipeline

This is because `contract!` in `ITensorGPU.jl` expects only `Index` objects for indices of dense arrays.

https://github.com/ITensor/ITensors.jl/blob/main/ITensorGPU/src/tensor/cudense.jl#L238