@shinaoka 2023-02-24
I have set up a local GPU environment. I will investigate the failure in ITensorGPU.