@shinaoka 2023-03-26
A minimum code to reproduce the failure in ITensorGPU.

```julia
using ITensors,  ITensorGPU

mi, mj, mk, ml = 1, 1, 2, 2
i = Index(mi, "i")
j = Index(mj, "j")
k = Index(mk, "k")
l = Index(ml, "l")

A = randomCuITensor(Float64, i, j, k, l)
U, S, V = svd(A, (j, l))

@assert cpu(U) * cpu(V) ≈ cpu(U * V) # OK
@assert cpu(S) * cpu(V) ≈ cpu(S * V) # Fail
```